refactor: change folder for controlled-select
